Moved FP Converters
Moved the fixed point conversion functions around for better organization.
