+++ vita49_2/vita49_2_packet_types.h
@@ -234,54 +234,4 @@ __vrt_api void vrt_set_payload_double(uint32_t *payload, size_t max_words, size_
  */
 __vrt_api int vrt_parse_cif_payload(const struct vrt_packet *pkt, struct vrt_cif_fields *cif);
 
-/**
- * @brief Converts a double to a 64-bit integer in 44.20 format (meaning the radix point is to the left of the first 20 bits).
- * 
- * @param value 
- * @return __vrt_api 
- */
-__vrt_api int64_t convert_to_44_20(double value);
-
-/**
- * @brief Converts a 64-bit integer in 44.20 format (meaning the radix point is to the left of the first 20 bits) back to a double.
- * 
- * @param value 
- * @return __vrt_api 
- */
-__vrt_api double convert_from_44_20(int64_t value);
-
-/**
- * @brief Converts a 16-bit FP-value to a 16-bit integer in 10.6 format (meaning the radix point is to the left of the first 6 bits).
- * 
- * @param value 
- * @return __vrt_api 
- */
-__vrt_api int16_t convert_to_10_6(float value);
-
-/**
- * @brief Converts a 16-bit integer in 10.6 format (meaning the radix point is to the left of the first 6 bits) back to a float.
- * 
- * @param value 
- * @return __vrt_api 
- */
-__vrt_api float convert_from_10_6(int16_t value);
-
-
-/**
- * @brief Converts a 16-bit FP-value to a 16-bit integer in 9.7 format (meaning the radix point is to the left of the first 7 bits).
- * 
- * @param value 
- * @return __vrt_api 
- */
-__vrt_api int16_t convert_to_9_7(float value);
-
-/**
- * @brief Converts a 16-bit integer in 9.7 format (meaning the radix point is to the left of the first 7 bits) back to a float.
- * 
- * @param value 
- * @return __vrt_api 
- */
-__vrt_api float convert_from_9_7(int16_t value);
-
-
 #endif /* __VITA49_PACKET_H__ */
